Codestyle Improvements.
