+++ Cauli/Florets/Mock/SwappedRecord.swift
@@ -74,98 +74,3 @@ extension SwappedRecord {
         return record
     }
 }
-
-internal struct SwappedURLRequest: Codable {
-    let url: URL
-    let cachePolicy: URLRequest.CachePolicy
-    let timeoutInterval: TimeInterval
-    let mainDocumentURL: URL?
-    let allowsCellularAccess: Bool
-    let httpMethod: String?
-    let allHTTPHeaderFields: [String: String]?
-    let bodyFilename: String?
-    let httpShouldHandleCookies: Bool
-    let httpShouldUsePipelining: Bool
-}
-
-extension SwappedURLRequest {
-    init(_ urlRequest: URLRequest, bodyFilepath: URL) {
-        self.url = urlRequest.url ?? URL(string: "unknown")!
-        self.cachePolicy = urlRequest.cachePolicy
-        self.timeoutInterval = urlRequest.timeoutInterval
-        self.mainDocumentURL = urlRequest.mainDocumentURL
-        self.allowsCellularAccess = urlRequest.allowsCellularAccess
-        self.httpMethod = urlRequest.httpMethod
-        self.allHTTPHeaderFields = urlRequest.allHTTPHeaderFields
-        self.httpShouldHandleCookies = urlRequest.httpShouldHandleCookies
-        self.httpShouldUsePipelining = urlRequest.httpShouldUsePipelining
-
-        if let body = urlRequest.httpBody,
-            (try? body.write(to: bodyFilepath)) != nil {
-            self.bodyFilename = bodyFilepath.lastPathComponent
-        } else {
-            self.bodyFilename = nil
-        }
-    }
-
-    func urlRequest(in folder: URL) -> URLRequest {
-        var request = URLRequest(url: url, cachePolicy: cachePolicy, timeoutInterval: timeoutInterval)
-        request.mainDocumentURL = mainDocumentURL
-        request.allowsCellularAccess = allowsCellularAccess
-        request.httpMethod = httpMethod
-        request.allHTTPHeaderFields = allHTTPHeaderFields
-        request.httpShouldHandleCookies = httpShouldHandleCookies
-        request.httpShouldUsePipelining = httpShouldUsePipelining
-
-        if let bodyFilename = bodyFilename,
-            let body = try? Data(contentsOf: folder.appendingPathComponent(bodyFilename)) {
-            request.httpBody = body
-        }
-
-        return request
-    }
-}
-
-internal struct SwappedResponse: Codable {
-    let dataFilename: String?
-    var urlResponse: URLResponse {
-        get {
-            return urlResponseRepresentable.urlResponse
-        }
-        set {
-            urlResponseRepresentable = URLResponseRepresentable(newValue)
-        }
-    }
-
-    init(_ urlResponse: URLResponse, dataFilename: String?) {
-        self.dataFilename = dataFilename
-        urlResponseRepresentable = URLResponseRepresentable(urlResponse)
-    }
-
-    var urlResponseRepresentable: URLResponseRepresentable
-}
-
-extension SwappedResponse {
-    init(_ response: Response, dataFilepath: URL) {
-        self.urlResponseRepresentable = URLResponseRepresentable(response.urlResponse)
-
-        if let data = response.data,
-            (try? data.write(to: dataFilepath)) != nil {
-            self.dataFilename = dataFilepath.lastPathComponent
-        } else {
-            self.dataFilename = nil
-        }
-    }
-
-    func response(in folder: URL) -> Response {
-        var response = Response(urlResponse, data: nil)
-        response.data = nil
-
-        if let dataFilename = dataFilename,
-            let data = try? Data(contentsOf: folder.appendingPathComponent(dataFilename)) {
-            response.data = data
-        }
-
-        return response
-    }
-}
