+++ Example/cauli-ios-example/cauli-ios-example/AppDelegate.swift
@@ -16,7 +16,7 @@ class AppDelegate: UIResponder, UIApplicationDelegate {
 
     func application(_ application: UIApplication, willFinishLaunchingWithOptions launchOptions: [UIApplication.LaunchOptionsKey : Any]? = nil) -> Bool {
         Cauli.customShared.run()
-        Cauli.mockFloret.mode = .mock
+        Cauli.mockFloret.mode = .record
         Cauli.mockFloret.enabled = true
         return true
     }
